add authentication middleware

+++ src/app/api/trainings/[id]/route.ts
@@ -19,7 +19,10 @@ export async function PATCH(req: Request, { params }: { params: { id: string } }
     return Response.json(training, { status: 200 })
 }
 
-export async function DELETE(_: Request, { params }: { params: { id: string } }) {
+export async function DELETE(req: Request, { params }: { params: { id: string } }) {
+    if (!req.headers.get("isAdmin")) {
+        return Response.json({ error: "Unauthorized" }, { status: 401 })
+    }
     await Training.findByIdAndDelete(params.id);
     return Response.json({}, { status: 200 });
 }
